feat: Add SpeakerDetailsForm component for speaker information management
- Introduced SpeakerDetailsForm to handle speaker details including name, title, bio, email, image upload, and social links.
- Removed unused ProfileError, ProfileImageUpload, ProfileLinksManager, ProfilePageClient, ProfileSkeleton, and SpeakerProfileForm components.
- Updated state management and validation for speaker details.
- Enhanced UI with appropriate error handling and loading states.
