+++ src/components/cfp/ProposalForm.tsx
@@ -1,12 +1,7 @@
 'use client'
 
 import { Conference } from '@/lib/conference/types'
-import {
-  getEmails,
-  postImage,
-  putEmail,
-  putProfile,
-} from '@/lib/profile/client'
+import { getEmails, putProfile } from '@/lib/profile/client'
 import { ProfileEmail } from '@/lib/profile/types'
 import { postProposal } from '@/lib/proposal'
 import {
@@ -22,27 +17,25 @@ import {
   ProposalInput,
   ProposalExisting,
 } from '@/lib/proposal/types'
-import { Flags, SpeakerInput, Speaker } from '@/lib/speaker/types'
+import { SpeakerInput, Speaker } from '@/lib/speaker/types'
 import { Topic } from '@/lib/topic/types'
-import { UserCircleIcon, XCircleIcon } from '@heroicons/react/24/solid'
+import { XCircleIcon } from '@heroicons/react/24/solid'
 import { PortableTextBlock } from '@portabletext/editor'
-import Image from 'next/image'
 import { redirect } from 'next/navigation'
 import { useEffect, useState } from 'react'
 import {
   Checkbox,
   Dropdown,
-  ErrorText,
   HelpText,
   Input,
-  LinkInput,
   Multiselect,
   Textarea,
-} from './Form'
-import { PortableTextEditor } from './PortableTextEditor'
-import { CoSpeakerSelector } from './CoSpeakerSelector'
+} from '../Form'
+import { PortableTextEditor } from '../PortableTextEditor'
+import { ProposalCoSpeaker } from './ProposalCoSpeaker'
 import { CoSpeakerInvitationMinimal } from '@/lib/cospeaker/types'
 import Link from 'next/link'
+import { SpeakerDetailsForm } from './SpeakerDetailsForm'
 
 export function ProposalForm({
   initialProposal,
@@ -189,7 +182,7 @@ export function ProposalForm({
           allowedFormats={allowedFormats}
         />
         <div className="border-b border-brand-frosted-steel pb-12">
-          <CoSpeakerSelector
+          <ProposalCoSpeaker
             selectedSpeakers={coSpeakers}
             onSpeakersChange={setCoSpeakers}
             format={proposal.format}
@@ -423,309 +416,6 @@ function ProposalDetailsForm({
               to submit your presentation.
             </HelpText>
           </Checkbox>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function SpeakerDetailsForm({
-  speaker,
-  setSpeaker,
-  email,
-  emails,
-}: {
-  speaker: SpeakerInput
-  setSpeaker: (speaker: SpeakerInput) => void
-  email: string
-  emails: ProfileEmail[]
-}) {
-  const [speakerName, setSpeakerName] = useState(speaker?.name ?? '')
-  const [speakerTitle, setSpeakerTitle] = useState(speaker?.title ?? '')
-  const [speakerBio, setSpeakerBio] = useState(speaker?.bio ?? '')
-  const [speakerEmail, setSpeakerEmail] = useState(email)
-  const [speakerImage, setSpeakerImage] = useState(speaker?.image ?? '')
-  const [speakerFlags, setSpeakerFlags] = useState(speaker?.flags ?? [])
-  const [speakerLinks, setSpeakerLinks] = useState(
-    speaker?.links?.length ? speaker.links : [''],
-  )
-
-  const [imageError, setImageError] = useState('')
-  const [isUploading, setIsUploading] = useState(false)
-
-  const emailOptions = new Map(
-    emails.map((email) => [email.email, email.email]),
-  )
-
-  function updateSpeakerFlag(flag: Flags, value: boolean) {
-    if (value) {
-      setSpeakerFlags([...speakerFlags, flag])
-    } else {
-      setSpeakerFlags(speakerFlags.filter((f) => f !== flag))
-    }
-  }
-
-  function updateSpeakerLink(i: number, val: string) {
-    setSpeakerLinks(
-      speakerLinks.map((link, index) => (index === i ? val : link)),
-    )
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  function addSpeakerLink(i: number) {
-    setSpeakerLinks([...speakerLinks, ''])
-  }
-
-  function removeSpeakerLink(i: number) {
-    const links = speakerLinks.filter((link, index) => index !== i)
-    if (links.length === 0) {
-      links.push('')
-    }
-    setSpeakerLinks(links)
-  }
-
-  async function imageUploadHandler(e: React.ChangeEvent<HTMLInputElement>) {
-    if (e.target.files && e.target.files.length > 0) {
-      setIsUploading(true)
-      const { error, image } = await postImage(e.target.files[0])
-
-      if (error) {
-        setImageError(error.message)
-      } else if (image) {
-        setSpeakerImage(image.image)
-      }
-
-      setIsUploading(false)
-    }
-  }
-
-  async function emailSelectHandler(email: string) {
-    const res = await putEmail(email)
-    if (res.error) {
-      // @TODO: display error message
-      console.error('Error updating email', res.error)
-    } else {
-      setSpeakerEmail(email)
-    }
-  }
-
-  useEffect(() => {
-    const links = speakerLinks.filter((link) => link.length > 0)
-
-    setSpeaker({
-      name: speakerName,
-      title: speakerTitle,
-      bio: speakerBio,
-      flags: speakerFlags,
-      links,
-    })
-  }, [
-    speakerName,
-    speakerTitle,
-    speakerBio,
-    speakerFlags,
-    speakerLinks,
-    setSpeaker,
-  ])
-
-  return (
-    <div className="border-b border-brand-frosted-steel pb-12">
-      <h2 className="font-space-grotesk text-lg leading-7 font-semibold text-brand-slate-gray">
-        Speaker Information
-      </h2>
-      <p className="font-inter mt-1 text-sm leading-6 text-brand-cloud-gray">
-        We need information about you as the speaker.
-      </p>
-
-      <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-        <div className="sm:col-span-4">
-          <Input
-            name="speaker_name"
-            label="Name"
-            value={speakerName}
-            setValue={setSpeakerName}
-          />
-        </div>
-
-        <div className="sm:col-span-4">
-          <Input
-            name="speaker_title"
-            label="Title or affiliation"
-            value={speakerTitle}
-            setValue={setSpeakerTitle}
-          />
-        </div>
-
-        <div className="col-span-full">
-          <Textarea
-            name="speaker_bio"
-            label="Bio"
-            rows={3}
-            value={speakerBio}
-            setValue={setSpeakerBio}
-          />
-          <HelpText>
-            This is what will be displayed to the audience on the conference
-            website. It should provide information about you as a speaker and
-            your expertise.
-          </HelpText>
-        </div>
-
-        <div className="sm:col-span-4">
-          <Dropdown
-            name="speaker_email"
-            label="Email address"
-            value={speakerEmail}
-            setValue={emailSelectHandler}
-            options={emailOptions}
-          />
-          <HelpText>
-            Your email address will not be displayed publicly. It will only be
-            used to contact you regarding your presentation.
-          </HelpText>
-        </div>
-
-        <div className="col-span-full">
-          <label
-            htmlFor="photo"
-            className="font-space-grotesk block text-sm leading-6 font-medium text-brand-slate-gray"
-          >
-            Photo
-          </label>
-          <div className="mt-2 flex items-center gap-x-3">
-            {speakerImage ? (
-              <Image
-                src={`${speakerImage}?w=96&h=96&fit=crop`}
-                alt="Speaker Image"
-                width={48}
-                height={48}
-                className="h-12 w-12 rounded-full"
-              />
-            ) : (
-              <UserCircleIcon
-                className="h-12 w-12 text-gray-300"
-                aria-hidden="true"
-              />
-            )}
-            <input
-              type="file"
-              id="photo"
-              className="sr-only"
-              accept="image/*"
-              onChange={imageUploadHandler}
-            />
-            {isUploading ? (
-              <div className="flex items-center gap-x-2">
-                <div className="h-5 w-5 animate-spin rounded-full border-t-2 border-b-2 border-brand-cloud-blue"></div>
-                <p className="font-inter text-sm leading-6 font-medium text-brand-slate-gray">
-                  Uploading...
-                </p>
-              </div>
-            ) : (
-              <label htmlFor="photo" className="cursor-pointer">
-                <span className="font-inter text-sm leading-6 font-medium text-brand-cloud-blue hover:text-brand-cloud-blue-hover">
-                  Upload Photo
-                </span>
-              </label>
-            )}
-          </div>
-          {imageError ? (
-            <ErrorText>{imageError}</ErrorText>
-          ) : (
-            <HelpText>
-              Your photo will be displayed on the conference website.
-            </HelpText>
-          )}
-        </div>
-
-        <div className="sm:col-span-4">
-          <fieldset>
-            <legend className="font-space-grotesk text-sm leading-6 font-semibold text-brand-slate-gray">
-              Social profiles and links
-            </legend>
-            <HelpText>
-              Provide links to your social profiles, personal website or other
-              relevant links you want to share with the audience.
-            </HelpText>
-            <div className="mt-6 space-y-6">
-              {speakerLinks.map((link, index) => (
-                <LinkInput
-                  index={index}
-                  key={`speaker_link_${index} `}
-                  name={`speaker_link_${index} `}
-                  value={link}
-                  update={updateSpeakerLink}
-                  remove={removeSpeakerLink}
-                  add={addSpeakerLink}
-                />
-              ))}
-            </div>
-          </fieldset>
-        </div>
-
-        <div className="col-span-full">
-          <fieldset>
-            <legend className="font-space-grotesk text-sm leading-6 font-semibold text-brand-slate-gray">
-              Speaker Details
-            </legend>
-            <div className="mt-6 space-y-6">
-              <Checkbox
-                name="local"
-                label="I am a local speaker"
-                value={speakerFlags.includes(Flags.localSpeaker)}
-                setValue={(value: boolean) =>
-                  updateSpeakerFlag(Flags.localSpeaker, value)
-                }
-              >
-                <HelpText>
-                  Please indicate if you are a local speaker to help us promote
-                  local talent.
-                </HelpText>
-              </Checkbox>
-
-              <Checkbox
-                name="first-time"
-                label="I am a first time speaker"
-                value={speakerFlags.includes(Flags.firstTimeSpeaker)}
-                setValue={(value: boolean) =>
-                  updateSpeakerFlag(Flags.firstTimeSpeaker, value)
-                }
-              >
-                <HelpText>
-                  We encourage new speakers to submit to this conference and
-                  will provide support and guidance if needed.
-                </HelpText>
-              </Checkbox>
-
-              <Checkbox
-                name="diverse"
-                label="I am from an underrepresented group"
-                value={speakerFlags.includes(Flags.diverseSpeaker)}
-                setValue={(value: boolean) =>
-                  updateSpeakerFlag(Flags.diverseSpeaker, value)
-                }
-              >
-                <HelpText>
-                  We are committed to increase diversity among our selected
-                  speakers.
-                </HelpText>
-              </Checkbox>
-
-              <Checkbox
-                name="requires-funding"
-                label="I require funding to attend the conference"
-                value={speakerFlags.includes(Flags.requiresTravelFunding)}
-                setValue={(value: boolean) =>
-                  updateSpeakerFlag(Flags.requiresTravelFunding, value)
-                }
-              >
-                <HelpText>
-                  If you require funding to attend the conference, please
-                  indicate it here.
-                </HelpText>
-              </Checkbox>
-            </div>
-          </fieldset>
         </div>
       </div>
     </div>
